+++ src/parts/printFileJobStartInfo.js
@@ -3,7 +3,5 @@
 const printer = require("../shared/printer");
 
 module.exports = context => {
-	printer.info(`File name ${context.fullFileName}`);
-	printer.info(`Target directory ${context.containingFolder}`);
-	printer.info(`Template to use ${context.template}`);
+	printer.debug(`Started File: ${context.fullFileName}`);
 };